bump to 2.0.0.
The only breaking change of note is that `svgo` is upgraded to 3.x. The only breaking changes are for users specifying `optimizeSvgs: true`. In this case, users may need to adjust the `svgoOptions` they are passing. Check out the `svgo` [release notes](https://github.com/svg/svgo/releases/tag/v3.0.0) for more information.